requirements.txt
asgiref==3.9.1
Django==5.2.4
django-js-asset==3.1.2
django-mptt==0.17.0
pillow==11.3.0
sqlparse==0.5.3
